requirements.txt
streamlit>=1.30.0
gradio>=4.0.0
anthropic>=0.40.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
